requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
line-bot-sdk>=3.5.0
googlemaps>=4.10.0
pydantic>=2.4.2
fastapi-mcp>=0.3.3
openai>=1.6.0
pyyaml>=6.0
requests>=2.31.0
